requirements.txt
# 核心依赖
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
openai>=1.0.0
datasets>=2.14.0,<4.0.0
networkx>=3.0
pydantic>=2.0.0

# 工具库
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
gdown>=4.6.0  # 用于下载 DocRED 数据
python-dotenv>=1.0.0  # 用于加载 .env 文件
spacy>=3.7.0

# 可选：itext2kg/ATOM 抽取器（如需启用 backend/python_service/config/modules.yaml 里的 itext2kg extractor）
# 参考：https://github.com/AuvaLab/itext2kg
itext2kg>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0

# Jupyter 支持（用于 notebook）
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 可选：用于可视化图结构
pyvis>=0.3.0
